pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solsat-poc-phr2"
version = "2.0.0"
description = "Solar Analysis API with Enhanced Features"
authors = [
    {name = "Solar Analysis Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.0",
    "numpy>=1.24.0",
    "shapely>=2.0.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.1"
]
requires-python = ">=3.9"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "httpx>=0.24.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | main
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]
skip_glob = ["main/*", "venv/*"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "main"
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "main/",
    "venv/",
    "build/"
]

[[tool.mypy.overrides]]
module = [
    "shapely.*",
    "folium.*",
    "matplotlib.*",
    "numpy.*",
    "PIL.*"
]
ignore_missing_imports = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
